Move unit tests to own package and update imports
