+++ jest.config.js
@@ -1,7 +1,7 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
-  setupFiles: ['./test/setup.ts'],
+  setupFiles: ['./test/unit-tests/setup.ts'],
   globals: {
     'ts-jest': {
         diagnostics: false
