+++ test/integration-tests/src/client-util.js
@@ -1,14 +1,11 @@
-require('dotenv').config();
+const configUtil = require('./config-util');
 const revai = require('revai-node-sdk');
 
-function getClientBaseUrl() {
-    return process.env.BASE_URL;
-}
 
 module.exports = {
     getClient: function (apiKey) {
         let client = new revai.RevAiApiClient(apiKey);
-        client.apiHandler.instance.defaults.baseURL = getClientBaseUrl();
+        client.apiHandler.instance.defaults.baseURL = configUtil.getBaseUrl();
         return client;
     }
 };
