+++ test/integration-tests/src/config-util.js
@@ -7,5 +7,8 @@ module.exports = {
     },
     getUserEmail: function() {
         return process.env.USER_EMAIL;
+    },
+    getBaseUrl: function() {
+        return process.env.BASE_URL;
     }
 }
