+++ test/integration-tests/test/account.test.js
@@ -1,11 +1,9 @@
 const clientUtil = require('../src/client-util');
-const userUtil = require('../src/user-util');
+const configUtil = require('../src/config-util');
 
 test('Should return email and balance', async () => {
-    const apiKey = userUtil.getApiKey();
-    let client = clientUtil.getClient(apiKey);
+    let client = clientUtil.getClient(configUtil.getApiKey());
     const account = await client.getAccount();
-    const userEmail = userUtil.getUserEmail();
-    expect(account.email).toBe(userEmail);
+    expect(account.email).toBe(configUtil.getUserEmail());
     expect(account.balance_seconds).not.toBeNull();
 });
