+++ test/unit-tests/api-client.spec.ts
@@ -1,11 +1,11 @@
-import { RevAiApiClient } from '../src/api-client';
-import { ApiRequestHandler } from '../src/api-request-handler';
-import { RevAiApiTranscript } from '../src/models/RevAiApiTranscript';
+import { RevAiApiClient } from '../../src/api-client';
+import { ApiRequestHandler } from '../../src/api-request-handler';
+import { RevAiApiTranscript } from '../../src/models/RevAiApiTranscript';
 import { objectToStream } from './testhelpers';
 const fs = require('fs');
 const FormData = require('form-data');
 
-jest.mock('../src/api-request-handler');
+jest.mock('../../src/api-request-handler');
 
 let sut: RevAiApiClient;
