+++ test/unit-tests/streaming-client.spec.ts
@@ -1,7 +1,7 @@
 import { WebSocketClient, WebSocketConnection } from 'websocket';
-import { AudioConfig } from '../src/models/streaming/AudioConfig';
-import { BufferedDuplex } from '../src/models/streaming/BufferedDuplex';
-import { RevAiStreamingClient } from '../src/streaming-client';
+import { AudioConfig } from '../../src/models/streaming/AudioConfig';
+import { BufferedDuplex } from '../../src/models/streaming/BufferedDuplex';
+import { RevAiStreamingClient } from '../../src/streaming-client';
 const fs = require('fs');
 const events = require('events');
